// FCAPROGRAMACION/FrontEnd/FCAPROGCAT006CW/src/app/models/paros.ts
export class Paros {
    tipotiempo: string;
    clavemaquina: string;
    fecha: string;
    horaini: string;
    horafin: string;
    tiempomin: number;
    comentario: string;
    estatus : boolean;  
    usuario: string;
    /*tipoparo : number;
    tipomaquina : string;
    clavemaquina : string;
    fechaini : string;
    horaini : string;
    fechafin : string;
    horafin : string;
    tiempomin : number;
    comentario: string;
    estatus : boolean    
    usuarioinsert: string;
    fechainsert : string;
    usuarioupdate: string;
    fechaupdate : string;
    usuariodelete: string;
    fechadelete : string;*/
}